--- src/transaction_reader_csv_xlsx.py
@@ -1,19 +1,0 @@
-from typing import Dict, List
-
-import pandas as pd
-
-
-def read_transactions_csv(file_path: str) -> List[Dict]:
-    """
-    Функция для считывания финансовых операций из CSV
-    """
-    csv_data = pd.read_csv(file_path, sep=";")
-    return csv_data.to_dict(orient="records")
-
-
-def read_transactions_xlsx(file_path: str) -> List[Dict]:
-    """
-    Функция для считывания финансовых операций из Excel
-    """
-    xlsx_data = pd.read_excel(file_path)
-    return xlsx_data.to_dict(orient="records")
