+++ tests/test_utils.py
@@ -1,9 +1,9 @@
 import json
-from unittest.mock import mock_open, patch
+from unittest.mock import MagicMock, mock_open, patch
 
 import pytest
 
-from src.utils import load_json_data
+from src.utils import load_json_data, read_transactions_csv, read_transactions_xlsx
 
 
 # Фикстуры для тестовых данных
@@ -122,3 +122,39 @@ def test_load_json_data_file_path_variations():
         for path in paths:
             result = load_json_data(path)
             assert result == test_data
+
+
+def test_read_transactions_csv():
+    mock_df = MagicMock()
+    mock_df.to_dict.return_value = [
+        {"date": "2024-01-01", "amount": 100, "description": "income"},
+        {"date": "2024-01-02", "amount": -50, "description": "expense"},
+    ]
+
+    with patch("pandas.read_csv", return_value=mock_df) as mock_read_csv:
+        result = read_transactions_csv("fake_path.csv")
+        mock_read_csv.assert_called_once_with("fake_path.csv", sep=";")
+        mock_df.to_dict.assert_called_once_with(orient="records")
+
+    assert isinstance(result, list)
+    assert len(result) == 2
+    assert result[0]["amount"] == 100
+    assert result[1]["description"] == "expense"
+
+
+def test_read_transactions_xlsx():
+    mock_df = MagicMock()
+    mock_df.to_dict.return_value = [
+        {"date": "2024-01-01", "amount": 200, "description": "salary"},
+        {"date": "2024-01-02", "amount": -80, "description": "bill"},
+    ]
+
+    with patch("pandas.read_excel", return_value=mock_df) as mock_read_excel:
+        result = read_transactions_xlsx("fake_path.xlsx")
+        mock_read_excel.assert_called_once_with("fake_path.xlsx")
+        mock_df.to_dict.assert_called_once_with(orient="records")
+
+    assert isinstance(result, list)
+    assert len(result) == 2
+    assert result[0]["amount"] == 200
+    assert result[1]["description"] == "bill"
